fix: Fix handling of resending messages during a disconnect.
The unit tests for this had a fatal flaw - they assumed that if the
connection dropped, we'd get nothing back for any in-flight messages.
This isn't true, though - we'd actually get back an error from amqplib
when the underlying connection fails.  This fixes the tests to reflect
this.  If we rely on amqplib to reject such messages, then moving all
messages from _unconfirmedMessages to _messages on a
reconnect now becomes superfluous.

I also reworked `_publishQueuedMessages()` to be more synchronous.
As it stood, I had to add a lot of pointless delays in my tests to make
sure that the `then` after publishing a message had time to run.

fix #152